iot/requirements.txt
paho-mqtt>=1.6.1
websockets>=12.0
aliyun-python-sdk-core>=2.13.0
aliyun-python-sdk-iot>=2.0.0
